Created validation for all the edit text fields.
Created validation for all the text fields.
Names will only take letters, periods, apostrophes, and dashes.
Number will take either 11 or 10 digits. The text fields is limited to 11.
Email will have the format of email@address.domain.

These do not prevent the entered value from showing up on the card.
If the validation requirements are not met, the text fields turn red and a message is shown below it.