Created Friend Code System
Updated the saveCard method on page C.
The information now saves to a document that is named after the user's unique fireAuth id.

Also created a textfield and button on page A.
This allows the user to add other user's ids to their collection.
Validation is implemented for this as well.

Still need to figure out how to access the fields as a string, so we can display the information.
Also need to implement delete card feature.